+++ package.json
@@ -12,8 +12,11 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.20.2",
+    "child_process": "^1.0.2",
     "essentia.js": "^0.1.3",
     "express": "^4.18.2",
+    "ffmpeg": "^0.0.4",
+    "fluent-ffmpeg": "^2.1.2",
     "multer": "^1.4.5-lts.1"
   }
 }
